refactor(domain): turn block into an immutable value type
Last of the value-types series. Merge `block_basis` into `block` as a
single class, drop all internal cache and the mutex (`total_inputs_`,
`base_size_`, `mutex_`). `serialized_size()` and `total_inputs()`
recompute on every call.

Also drop default parameters from the public API touched by this diff
(callers must spell every argument explicitly):
- `from_data(byte_reader&)` — `wire` parameter removed entirely;
  block serialization is always canonical encoding, so the flag had
  exactly one valid value.
- `subsidy(size_t, bool retarget)` — default removed.
- `total_inputs(bool with_coinbase)` (member + non-member free fn) —
  default removed.

Two latent bugs surfaced and fixed:
- `chain::block::to_data(true)` / `to_data(false)` in database/test
  call sites were silently resolving to `to_data(size_t)` via implicit
  `bool→size_t` conversion, producing a `serialized_size = 1` argument
  that would have tripped the `KTH_ASSERT(data.size() == size)` in
  debug builds. Replaced with `to_data()`.
- C-API `kth_chain_block_construct_from_data` carried a vestigial
  `kth_bool_t wire` parameter mirroring the C++ default; removed from
  both the header and tests. Same for `kth_chain_block_to_data` which
  took a redundant `kth_size_t serialized_size`.

Net diff: -352 LOC.
