+++ src/database/include/kth/database/databases/reorg_database.ipp
@@ -55,7 +55,7 @@ result_code internal_database_basis<Clock>::insert_reorg_pool(uint32_t height, K
 template <typename Clock>
 result_code internal_database_basis<Clock>::push_block_reorg(domain::chain::block const& block, uint32_t height, KTH_DB_txn* db_txn) {
 
-    auto valuearr = block.to_data(false);               //TODO(fernando): podría estar afuera de la DBTx
+    auto valuearr = block.to_data();               //TODO(fernando): podría estar afuera de la DBTx
     auto key = kth_db_make_value(sizeof(height), &height);              //TODO(fernando): podría estar afuera de la DBTx
     auto value = kth_db_make_value(valuearr.size(), valuearr.data());   //TODO(fernando): podría estar afuera de la DBTx
