+++ src/domain/include/kth/domain/chain/block.hpp
@@ -5,15 +5,14 @@
 #ifndef KTH_DOMAIN_CHAIN_BLOCK_HPP
 #define KTH_DOMAIN_CHAIN_BLOCK_HPP
 
+#include <algorithm>
 #include <cstddef>
 #include <cstdint>
 #include <istream>
 #include <memory>
-#include <optional>
 #include <string>
 #include <vector>
 
-#include <kth/domain/chain/block_basis.hpp>
 #include <kth/domain/chain/chain_state.hpp>
 #include <kth/domain/chain/header.hpp>
 #include <kth/domain/chain/transaction.hpp>
@@ -29,12 +28,13 @@
 #include <kth/infrastructure/utility/container_sink.hpp>
 #include <kth/infrastructure/utility/data.hpp>
 #include <kth/infrastructure/utility/reader.hpp>
-#include <kth/infrastructure/utility/thread.hpp>
 #include <kth/infrastructure/utility/writer.hpp>
 
 namespace kth::domain::chain {
 
-struct KD_API block : block_basis {
+using indexes = std::vector<size_t>;
+
+struct KD_API block {
 public:
     using list = std::vector<block>;
     using indexes = std::vector<size_t>;
@@ -68,42 +68,62 @@ public:
     block(chain::header const& header, transaction::list&& transactions);
     block(chain::header const& header, transaction::list const& transactions);
 
-    explicit
-    block(block_basis const& basis)
-        : block_basis(basis)
-    {}
+    // Operators.
+    //-------------------------------------------------------------------------
 
-    explicit
-    block(block_basis&& basis)
-        : block_basis(std::move(basis))
-    {}
-
-    //Note(kth): cannot be defaulted because of the mutex data member.
-    block(block const& x);
-    block(block&& x) noexcept;
-    /// This class is move assignable and copy assignable.
-    block& operator=(block const& x);
-    block& operator=(block&& x) noexcept;
+    bool operator==(block const& x) const;
+    bool operator!=(block const& x) const;
 
     // Deserialization.
     //-------------------------------------------------------------------------
+
     static
-    expect<block> from_data(byte_reader& reader, bool wire = true);
+    expect<block> from_data(byte_reader& reader);
+
+    [[nodiscard]]
+    bool is_valid() const;
 
     // Serialization.
     //-------------------------------------------------------------------------
 
-    using block_basis::to_data;
     data_chunk to_data() const;
+    void to_data(data_sink& stream) const;
 
-    // Properties (size, accessors, cache).
+    template <typename W>
+    void to_data(W& sink) const {
+        header_.to_data(sink, true);
+        sink.write_size_little_endian(transactions_.size());
+        auto const to = [&sink](transaction const& tx) {
+            tx.to_data(sink, true);
+        };
+        std::for_each(transactions_.begin(), transactions_.end(), to);
+    }
+
+    [[nodiscard]]
+    hash_list to_hashes() const;
+
+    // Properties (size, accessors).
     //-------------------------------------------------------------------------
 
     size_t serialized_size() const;
 
-    // void set_header(chain::header const& value);
+    chain::header& header();
+
+    [[nodiscard]]
+    chain::header const& header() const;
+
+    void set_header(chain::header const& value);
+
+    transaction::list& transactions();
+
+    [[nodiscard]]
+    transaction::list const& transactions() const;
+
     void set_transactions(transaction::list const& value);
     void set_transactions(transaction::list&& value);
+
+    [[nodiscard]]
+    hash_digest hash() const;
 
     // Utilities.
     //-------------------------------------------------------------------------
@@ -137,11 +157,62 @@ public:
     // Validation.
     //-------------------------------------------------------------------------
 
+    static
+    uint64_t subsidy(size_t height, bool retarget);
+
+    static
+    uint256_t proof(uint32_t bits);
+
+    [[nodiscard]]
+    uint64_t fees() const;
+
+    [[nodiscard]]
+    uint64_t claim() const;
+
+    [[nodiscard]]
+    uint64_t reward(size_t height) const;
+
+    [[nodiscard]]
+    uint256_t proof() const;
+
+    [[nodiscard]]
+    hash_digest generate_merkle_root() const;
+
+    [[nodiscard]]
     size_t signature_operations() const;
 
-    using block_basis::signature_operations;
+    [[nodiscard]]
+    size_t signature_operations(bool bip16, bool bip141) const;
 
-    size_t total_inputs(bool with_coinbase = true) const;
+    [[nodiscard]]
+    size_t total_inputs(bool with_coinbase) const;
+
+    [[nodiscard]]
+    bool is_extra_coinbases() const;
+
+    [[nodiscard]]
+    bool is_final(size_t height, uint32_t block_time) const;
+
+    [[nodiscard]]
+    bool is_distinct_transaction_set() const;
+
+    [[nodiscard]]
+    bool is_valid_coinbase_claim(size_t height) const;
+
+    [[nodiscard]]
+    bool is_valid_coinbase_script(size_t height) const;
+
+    [[nodiscard]]
+    bool is_forward_reference() const;
+
+    [[nodiscard]]
+    bool is_canonical_ordered() const;
+
+    [[nodiscard]]
+    bool is_internal_double_spend() const;
+
+    [[nodiscard]]
+    bool is_valid_merkle_root() const;
 
     code check() const;
 
@@ -149,19 +220,33 @@ public:
     /// Use this when headers have already been validated during header sync.
     code check_body() const;
 
+    [[nodiscard]]
+    code check_transactions() const;
+
     // THIS IS FOR LIBRARY USE ONLY, DO NOT CREATE A DEPENDENCY ON IT.
     mutable validation_t validation{};
 
-private:
-    mutable std::optional<size_t> total_inputs_;
-    mutable std::optional<size_t> base_size_;   // total size
+// protected:
+    void reset();
 
-#if ! defined(__EMSCRIPTEN__)
-    mutable upgrade_mutex mutex_;
-#else
-    mutable shared_mutex mutex_;
-#endif
+    [[nodiscard]]
+    size_t non_coinbase_input_count() const;
+
+private:
+    chain::header header_;
+    transaction::list transactions_;
 };
+
+// Non-member functions.
+//-------------------------------------------------------------------------
+
+size_t locator_size(size_t top);
+
+indexes locator_heights(size_t top);
+
+size_t total_inputs(block const& blk, bool with_coinbase);
+
+size_t serialized_size(block const& blk);
 
 } // namespace kth::domain::chain
