+++ src/domain/src/message/block.cpp
@@ -88,7 +88,7 @@ expect<block> block::from_data(byte_reader& reader, uint32_t /*version*/) {
 // Witness is always serialized if present.
 // NOTE: Witness on BCH is dissabled on the chain::block class
 data_chunk block::to_data(uint32_t /*unused*/) const {
-    return chain::block::to_data(true);
+    return chain::block::to_data();
 }
 
 void block::to_data(uint32_t /*version*/, data_sink& stream) const {
